Extract `IsStudentPermission` to `permissions` module
