+++ odin/education/apis/courses.py
@@ -1,24 +1,14 @@
 from rest_framework import serializers
 from rest_framework.generics import ListAPIView
-from rest_framework.permissions import BasePermission
 
 from odin.authentication.apis.permissions import JSONWebTokenAuthenticationMixin
 
-from odin.education.models import Course, Student
+from odin.education.models import Course
+
+from .permissions import IsStudentPermission
 
 
 class StudentCoursesApi(JSONWebTokenAuthenticationMixin, ListAPIView):
-    class IsStudentPermission(BasePermission):
-        def has_permission(self, request, view):
-            user = request.user
-            student = user.downcast(Student)
-
-            if student is not None:
-                view.student = student
-                return True
-
-            return False
-
     class Serializer(serializers.ModelSerializer):
         class Meta:
             model = Course
@@ -32,7 +22,7 @@ class StudentCoursesApi(JSONWebTokenAuthenticationMixin, ListAPIView):
     serializer_class = Serializer
 
     def get_permissions(self):
-        return super().get_permissions() + [self.IsStudentPermission()]
+        return super().get_permissions() + [IsStudentPermission()]
 
     def get_queryset(self):
         return Course.objects\
